update toolchain, minors (#16)

+++ shared/src/validator.rs
@@ -32,9 +32,8 @@ impl Validator {
             namada_core::address::gen_established_address("namada-indexer");
         let voting_power = (1000..10000000).fake::<u64>().to_string();
         let max_commission =
-            ((0..100).fake::<u64>() as f64 / 100 as f64).to_string();
-        let commission =
-            ((0..100).fake::<u64>() as f64 / 100 as f64).to_string();
+            ((0..100).fake::<u64>() as f64 / 100_f64).to_string();
+        let commission = ((0..100).fake::<u64>() as f64 / 100_f64).to_string();
         let email: String = SafeEmail().fake();
         let description: Option<String> = CatchPhase().fake();
         let website: Option<String> = Some(format!(
